Lumen: Change core value type from f32 to f64
I initially assumed f32, but in the course of writing the Lux
parser, I switched to f64, so replicating that change in Lumen.

+++ lumen/src/dmx.rs
@@ -8,7 +8,7 @@ impl Dmx {
         Self(value)
     }
 
-    pub fn from_factor(factor: f32) -> Self {
+    pub fn from_factor(factor: f64) -> Self {
         Self((255.0 * factor).round() as u8)
     }
 }
